+++ applications/fishing-map/src/features/editor/editor.hooks.ts
@@ -1,7 +1,7 @@
 import { useEffect, useCallback, useRef } from 'react'
 import { useSelector, useDispatch } from 'react-redux'
 import { selectEditorActive, toggleEditorMenu } from 'features/editor/editor.slice'
-import { isGFWUser } from 'features/user/user.slice'
+import { selectUserWorkspaceEditPermissions } from 'features/user/user.selectors'
 
 type EditorMenu = {
   editorActive: boolean
@@ -10,7 +10,7 @@ type EditorMenu = {
 
 const useEditorMenu = (): EditorMenu => {
   const dispatch = useDispatch()
-  const gfwUser = useSelector(isGFWUser)
+  const hasEditPermission = useSelector(selectUserWorkspaceEditPermissions)
   const numTimesEditorKeyDown = useRef(0)
   const editorKeyDownInterval = useRef<number>(0)
 
@@ -36,13 +36,13 @@ const useEditorMenu = (): EditorMenu => {
   )
 
   useEffect(() => {
-    if (gfwUser) {
+    if (hasEditPermission) {
       document.addEventListener('keydown', onKeyDown)
     }
     return () => {
       document.removeEventListener('keydown', onKeyDown)
     }
-  }, [gfwUser, onKeyDown])
+  }, [hasEditPermission, onKeyDown])
 
   const editorActive = useSelector(selectEditorActive)
   return { editorActive, dispatchToggleEditorMenu }