+++ applications/fishing-map/src/features/user/user.slice.ts
@@ -21,7 +21,6 @@ const initialState: UserState = {
 }
 
 export const GUEST_USER_TYPE = 'guest'
-export const ADMIN_GROUP_ID = 'Admin-group'
 export const GFW_GROUP_ID = 'GFW'
 
 export const fetchGuestUser = async () => {
@@ -91,7 +90,6 @@ const userSlice = createSlice({
 export const selectUserData = (state: RootState) => state.user.data
 export const selectUserStatus = (state: RootState) => state.user.status
 export const selectUserLogged = (state: RootState) => state.user.logged
-export const isAdminUser = (state: RootState) => state.user.data?.groups.includes(ADMIN_GROUP_ID)
 export const isGFWUser = (state: RootState) => state.user.data?.groups.includes(GFW_GROUP_ID)
 
 export default userSlice.reducer