+++ packages/layer-composer/src/generators/basemap/basemap-layers.ts
@@ -15,6 +15,16 @@ export const layers: Dictionary<Layer[]> = {
       },
     },
   ],
+  [BasemapType.Labels]: [
+    {
+      type: 'raster',
+      id: BasemapType.Labels,
+      source: BasemapType.Labels,
+      metadata: {
+        group: Group.BasemapForeground,
+      },
+    },
+  ],
   [BasemapType.Default]: [
     {
       id: 'bathymetry',
@@ -87,9 +97,17 @@ export const sources: Dictionary<Record<string, Partial<any>>> = {
   [BasemapType.Satellite]: {
     satellite: {
       tiles: [
-        'https://mt1.google.com/vt/lyrs=s&x={x}&y={y}&z={z}',
-        'https://mt2.google.com/vt/lyrs=s&x={x}&y={y}&z={z}',
-        'https://mt3.google.com/vt/lyrs=s&x={x}&y={y}&z={z}',
+        'https://gtiles.api.dev.globalfishingwatch.org/v1/tileset/sat/tile?x={x}&y={y}&z={z}',
+      ],
+      type: 'raster',
+      tileSize: 256,
+      attribution: 'Google',
+    },
+  },
+  [BasemapType.Labels]: {
+    labels: {
+      tiles: [
+        'https://gtiles.api.dev.globalfishingwatch.org/v1/tileset/nslabels/tile?x={x}&y={y}&z={z}',
       ],
       type: 'raster',
       tileSize: 256,