+++ applications/vessel-history/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @globalfishingwatchapp/vessel-history
 
+## 0.0.85
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+- Updated dependencies [9aa50a38]
+  - @globalfishingwatch/dataviews-client@11.13.1
+  - @globalfishingwatch/api-types@4.8.0
+
+## 0.0.84
+
+### Patch Changes
+
+- Updated dependencies [2c4a5386]
+  - @globalfishingwatch/dataviews-client@11.13.0
+  - @globalfishingwatch/layer-composer@8.14.0
+
+## 0.0.83
+
+### Patch Changes
+
+- Updated dependencies [1f6ccdf4]
+  - @globalfishingwatch/layer-composer@8.13.0
+
 ## 0.0.82
 
 ### Patch Changes