+++ packages/layer-composer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @globalfishingwatch/layer-composer
 
+## 8.14.0
+
+### Minor Changes
+
+- 2c4a5386: Basemap layer now supports a labels option
+
+## 8.13.0
+
+### Minor Changes
+
+- 1f6ccdf4: Customize inactive icon size and name prefix on events
+
 ## 8.12.1
 
 ### Patch Changes